+++ src/main/java/com/rmnnorbert/InquireNet/dto/delete/DeleteRequestDTO.java
@@ -1,4 +1,4 @@
 package com.rmnnorbert.InquireNet.dto.delete;
 
-public record DeleteRequestDTO (int userId, int targetId){
+public record DeleteRequestDTO (long userId, long targetId){
 }
