+++ src/main/java/com/rmnnorbert/InquireNet/dto/question/NewQuestionDTO.java
@@ -1,4 +1,4 @@
 package com.rmnnorbert.InquireNet.dto.question;
 
-public record NewQuestionDTO (String title, String description, Integer userID){
+public record NewQuestionDTO (String title, String description, long userID){
 }
